progress bar supports float value internally

+++ src/main/java/com/gtnewhorizons/wdmla/api/ui/style/IPanelStyle.java
@@ -12,5 +12,5 @@ public interface IPanelStyle {
     /**
      * @return the space pixel between each panel elements
      */
-    int getSpacing();
+    float getSpacing();
 }
