+++ src/main/java/com/gtnewhorizons/wdmla/api/ui/style/IRectStyle.java
@@ -22,5 +22,5 @@ public interface IRectStyle {
      */
     int getBorderColor();
 
-    int getBorderThickness();
+    float getBorderThickness();
 }
